+++ tool/dart_skills_lint/lib/src/models/custom_rule_parameters.dart
@@ -2,30 +2,26 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-/// A wrapper around raw rule options parameters.
+/// A wrapper around raw rule parameters.
 ///
 /// Prevents exposing raw [Map] APIs directly inside rule logic, and provides
 /// standard lookups and properties for rule configuration parameters.
-class CustomRuleOptions {
-  CustomRuleOptions(Map<String, dynamic> params)
+class CustomRuleParameters {
+  /// Creates a new configuration with the provided [params].
+  CustomRuleParameters(Map<String, dynamic> params)
     : params = Map<String, dynamic>.unmodifiable(params);
 
   /// The underlying map containing the parameters.
   final Map<String, dynamic> params;
 
-  /// Whether the configuration contains no parameters.
   bool get isEmpty => params.isEmpty;
 
-  /// Whether the configuration contains parameters.
   bool get isNotEmpty => params.isNotEmpty;
 
-  /// Retrieves the value of the parameter associated with [key].
   Object? operator [](String key) => params[key];
 
-  /// The parameter keys defined in this configuration.
   Iterable<String> get keys => params.keys;
 
-  /// Whether the configuration contains the specified [key].
   bool containsKey(String key) => params.containsKey(key);
 
   /// Retrieves the value of the parameter associated with [key] as a [String].